+++ designer_v2/lib/features/design/interventions/interventions_form_view.dart
@@ -72,7 +72,7 @@ class StudyDesignInterventionsFormView extends StudyDesignPageWidget {
                         hideLeadingTrailingWhenEmpty: true,
                         rowPrefix: (context, viewModel, rowIdx) =>
                             interventionPrefix(rowIdx, theme),
-                        reorderable: true,
+                        reorderable: !formViewModel.isReadonly,
                         onReorder: (oldIndex, newIndex) {
                           if (newIndex > oldIndex) {
                             newIndex -= 1;