+++ designer_v2/lib/localization/app_localizations_de.dart
@@ -726,7 +726,7 @@ class AppLocalizationsDe extends AppLocalizations {
   String get fitbit_question_title => 'Fitbit';
 
   @override
-  String get fitbit_question_type_empty => 'No Fitbit data available';
+  String get fitbit_question_type_empty => 'Keine Fitbit Daten verfügbar';
 
   @override
   String get navlink_question_visibility_logic => 'Sichtbarkeit';